Try to avoid errors in side thread

+++ jobmonitor/run.py
@@ -164,25 +164,25 @@ def main():
     )
 
     def side_thread_fn():
-        # Update the worker's heartbeat
-        try:
-            update_job(
-                job_id,
-                {
-                    "last_heartbeat_time": datetime.datetime.utcnow(),
-                    f"workers.{rank}.last_heartbeat_time": datetime.datetime.utcnow(),
-                },
-            )
-            # Check the status of the job and if we need to self-destruct
-            res = mongo.job.find_one({"_id": ObjectId(job_id)}, {"status": 1})
-            if res is None or res["status"] not in ["SCHEDULED", "RUNNING", "FINISHED"]:
-                status = res["status"] if res is not None else "DELETED"
-                print(f"Job status changed to {status}. This worker will self-destruct.")
-                os.kill(os.getpid(), signal.SIGUSR1)
-        except Exception:
-            import traceback
+        global GLOBAL_IS_EXITING
+        if GLOBAL_IS_EXITING:
+            return
+        # Check the status of the job and if we need to self-destruct
+        res = mongo.job.find_one({"_id": ObjectId(job_id)}, {"status": 1})
+        if res is None or res["status"] not in ["SCHEDULED", "RUNNING", "FINISHED"]:
+            status = res["status"] if res is not None else "DELETED"
+            print(f"Job status changed to {status}. This worker will self-destruct.")
+            os.kill(os.getpid(), signal.SIGUSR1)
+            return
 
-            print(traceback.format_exc())
+        # Update the worker's heartbeat
+        update_job(
+            job_id,
+            {
+                "last_heartbeat_time": datetime.datetime.utcnow(),
+                f"workers.{rank}.last_heartbeat_time": datetime.datetime.utcnow(),
+            },
+        )
 
     # Start sending regular heartbeat updates to the db
     # and check whether the job isn't getting canceled
@@ -274,7 +274,12 @@ class ExitCommand(Exception):
     pass
 
 
+GLOBAL_IS_EXITING = False
+
+
 def signal_handler(signal, frame):
+    global GLOBAL_IS_EXITING
+    GLOBAL_IS_EXITING = True
     raise ExitCommand()
